+++ main/src/ca/uwaterloo/flix/language/errors/ResolutionError.scala
@@ -16,6 +16,7 @@
 
 package ca.uwaterloo.flix.language.errors
 
+import ca.uwaterloo.flix.language.CompilationError
 import ca.uwaterloo.flix.language.ast.{Name, SourceInput, SourceLocation}
 import ca.uwaterloo.flix.util.vt.VirtualString._
 import ca.uwaterloo.flix.util.vt.VirtualTerminal
@@ -23,8 +24,7 @@ import ca.uwaterloo.flix.util.vt.VirtualTerminal
 /**
   * A common super-type for resolution errors.
   */
-sealed trait ResolutionError extends TypeError {
-  // TODO: Should extend CompilationError
+sealed trait ResolutionError extends CompilationError {
   val kind = "Resolution Error"
 }
 
@@ -73,25 +73,6 @@ object ResolutionError {
         vt << Code(l, "tag is defined in this enum.") << NewLine
       }
       vt << Underline("Tip:") << " Prefix the tag with the enum name." << NewLine
-    }
-  }
-
-  /**
-    * Undefined Attribute Error.
-    *
-    * @param attribute the attribute name.
-    * @param loc       the location where the error occurred.
-    */
-  case class UndefinedAttribute(table: String, attribute: String, loc: SourceLocation) extends ResolutionError {
-    val source: SourceInput = loc.source
-    val message: VirtualTerminal = {
-      val vt = new VirtualTerminal
-      vt << Line(kind, source.format) << NewLine
-      vt << ">> Undefined attribute '" << Red(attribute) << "' in table '" << Cyan(table) << "'." << NewLine
-      vt << NewLine
-      vt << Code(loc, "attribute not found.") << NewLine
-      vt << NewLine
-      vt << Underline("Tip:") << " Possible typo or non-existent attribute?" << NewLine
     }
   }
